tests/test_chatbot_integration.py
```python
import unittest
from unittest.mock import patch
from src import chatbot_integration

class TestChatbotIntegration(unittest.TestCase):

    @patch('src.chatbot_integration.requests.post')
    def test_integrateChat2Code(self, mock_post):
        # Mocking the API response
        mock_post.return_value.status_code = 200
        mock_post.return_value.json.return_value = {
            'status': 'success',
            'message': 'Chat2Code bot integrated successfully.'
        }

        response = chatbot_integration.integrateChat2Code('test_message')
        self.assertEqual(response['status'], 'success')
        self.assertEqual(response['message'], 'Chat2Code bot integrated successfully.')

    @patch('src.chatbot_integration.requests.post')
    def test_integrateChat2Code_failure(self, mock_post):
        # Mocking the API response
        mock_post.return_value.status_code = 400
        mock_post.return_value.json.return_value = {
            'status': 'failure',
            'message': 'Chat2Code bot integration failed.'
        }

        response = chatbot_integration.integrateChat2Code('test_message')
        self.assertEqual(response['status'], 'failure')
        self.assertEqual(response['message'], 'Chat2Code bot integration failed.')

if __name__ == '__main__':
    unittest.main()
```